discogs: make the script tougher against weird track numbering
http://www.discogs.com/release/73531 was leading to an error:
when it tries to parse track 3 position (AA1) it thinks it's a new disc and increase
releaseNumber.
Of course there is no matching release format, so it generates an exception..

http://www.discogs.com/Undertone-Sexy-Thing/release/1009184 also leads to the same exception, somehow for a similar reason (track numbering has the usual scheme),
but the real fix is on the discogs side for this one, tracks 7.1-7.6 should be track 7 with sub-tracks.
Just make the script tougher here again, it just tries to import tracks as they are.