qobuz_importer.user.js: Support for classical releases.
This will use composer as artist on various artists/composers releases.
Refactored code to functions as using composers as well as artists.
Added function for guessing various artists, as the hardcoded are
likely label based and are hard to keep updated.

Also, fixing a bug in single artist use case on a track.