Spotify importer: spotify_importer.user.js This piggybacks on the official web-player's API requests to parse information. This script is based on the qobuz_importer I modified some time ago.
From my testing, it seems the Spotify data is a bit crappy. There are
release dates that belong to the original releases. All the performers
seem to be 'artist', so featured and classical releases need cleaning when
importing. No language info.
The API limit seems to be 50 tracks - not tested.
I did not find multi-disc releases yet, so they are not tested.