mb_1200px_caa: New userscript!
This script makes the release/*/cover-art pages use the 1200px
“thumbnail” images for the pop‐up/preview instead of the original image,
which can speed up loading times significantly for large scans, but
should still be high enough definition for most text to be legible.

Additionally it adds a “1200px” link to the “All sizes” lists. This will
likely be unneeded as soon as CAA-88[1] is resolved, but until then it
is really nifty to have handy.

Note: A few (older) images may not be available in 1200px.

[1] https://tickets.metabrainz.org/browse/CAA-88